BUG: wigner_cmap not using int index (#703)
* faster rcm

* update norms

* Remove rcm bucky test

the use of int_argsort makes the exact perm array platform dependent.

* BUG: wigner_cmap not using int index

+++ qutip/matplotlib_utilities.py
@@ -106,8 +106,8 @@ def wigner_cmap(W, levels=1024, shift=0, max_color='#09224F',
     bounds = [W.min(), W.max()]
     # create empty array for RGBA colors
     adjust_RGBA = np.hstack((np.zeros((levels, 3)), np.ones((levels, 1))))
-    zero_pos = np.round(levels * np.abs(shift - bounds[0])
-                        / (bounds[1] - bounds[0]))
+    zero_pos = int(np.round(levels * np.abs(shift - bounds[0])
+                        / (bounds[1] - bounds[0])))
     num_pos = levels - zero_pos
     num_neg = zero_pos - 1
     # set zero values to mid_color
